Begin supporting static analysis of OLE and XML Office files. This commit heavily relies on Decalage's oletools code.
Supported files (via olevba.py):
- Word 97-2003 (.doc, .dot), Word 2007+ (.docm, .dotm)
 - Excel 97-2003 (.xls), Excel 2007+ (.xlsm, .xlsb)
 - PowerPoint 2007+ (.pptm, .ppsm)

We attempt to harvest metadata from OLE structures and attempt to extract VBA macros from all supported files types. VBA code is then processed and deobfuscated (best effort) so that we can parse for IOC's and suspicious keywords. This data is then added to the Django interface in the static analysis tab.

I reused a template tag that was with pdf analysis. It has now been migrated to a more appropriate tag file.